refactor: 픽스처 패키지를 별도로 두지 않고 도메인으로 이
