+++ src/test/java/nextstep/subway/favorite/FavoriteAcceptanceTest.java
@@ -2,7 +2,15 @@ package nextstep.subway.favorite;
 
 import nextstep.subway.AcceptanceTest;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 
 @DisplayName("즐겨찾기 관련 기능")
 public class FavoriteAcceptanceTest extends AcceptanceTest {
+    @DisplayName("즐겨찾기를 관리한다.")
+    @Test
+    void manageFavorite() {
+
+    }
+
+
 }
