+++ src/test/java/nextstep/subway/member/domain/MemberFixture.java
@@ -1,0 +1,5 @@
+package nextstep.subway.member.domain;
+
+public class MemberFixture {
+
+}
