+++ src/test/java/nextstep/subway/path/service/PathServiceTest.java
@@ -24,7 +24,7 @@ import java.util.stream.Stream;
 import static java.util.Arrays.asList;
 import static nextstep.subway.line.fixture.SectionFixture.전체_구간;
 import static nextstep.subway.station.dto.StationResponse.of;
-import static nextstep.subway.station.fixture.StationFixture.*;
+import static nextstep.subway.station.domain.StationFixture.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 import static org.mockito.Mockito.when;
